% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BFI_Extraversion}
\alias{BFI_Extraversion}
\title{Big Five Inventory (Full Dataset)}
\format{
A data frame with 57 columns:
\describe{
  \item{ID}{Participant identifier}
  \item{age}{Age in years}
  \item{gender}{Gender (coded)}
  \item{country}{Country of origin}
  \item{source}{Data collection source}
  \item{E1}{Extraversion item 1 (1--5 scale)}
  \item{E2}{Extraversion item 2, reverse-keyed (1--5 scale)}
  \item{E3}{Extraversion item 3 (1--5 scale)}
  \item{E4}{Extraversion item 4, reverse-keyed (1--5 scale)}
  \item{E5}{Extraversion item 5 (1--5 scale)}
  \item{E6}{Extraversion item 6, reverse-keyed (1--5 scale)}
  \item{E7}{Extraversion item 7 (1--5 scale)}
  \item{E8}{Extraversion item 8, reverse-keyed (1--5 scale)}
  \item{E9}{Extraversion item 9 (1--5 scale)}
  \item{E10}{Extraversion item 10, reverse-keyed (1--5 scale)}
  \item{E_mean}{Mean of Extraversion items}
  \item{A1}{Agreeableness item 1 (1--5 scale)}
  \item{A2}{Agreeableness item 2 (1--5 scale)}
  \item{A3}{Agreeableness item 3 (1--5 scale)}
  \item{A4}{Agreeableness item 4 (1--5 scale)}
  \item{A5}{Agreeableness item 5 (1--5 scale)}
  \item{A6}{Agreeableness item 6 (1--5 scale)}
  \item{A7}{Agreeableness item 7 (1--5 scale)}
  \item{A8}{Agreeableness item 8 (1--5 scale)}
  \item{A9}{Agreeableness item 9 (1--5 scale)}
  \item{A10}{Agreeableness item 10 (1--5 scale)}
  \item{A_mean}{Mean of Agreeableness items}
  \item{C1}{Conscientiousness item 1 (1--5 scale)}
  \item{C2}{Conscientiousness item 2 (1--5 scale)}
  \item{C3}{Conscientiousness item 3 (1--5 scale)}
  \item{C4}{Conscientiousness item 4 (1--5 scale)}
  \item{C5}{Conscientiousness item 5 (1--5 scale)}
  \item{C6}{Conscientiousness item 6 (1--5 scale)}
  \item{C7}{Conscientiousness item 7 (1--5 scale)}
  \item{C8}{Conscientiousness item 8 (1--5 scale)}
  \item{C9}{Conscientiousness item 9 (1--5 scale)}
  \item{C10}{Conscientiousness item 10 (1--5 scale)}
  \item{C_mean}{Mean of Conscientiousness items}
  \item{N1}{Neuroticism item 1 (1--5 scale)}
  \item{N2}{Neuroticism item 2 (1--5 scale)}
  \item{N3}{Neuroticism item 3 (1--5 scale)}
  \item{N4}{Neuroticism item 4 (1--5 scale)}
  \item{N5}{Neuroticism item 5 (1--5 scale)}
  \item{N6}{Neuroticism item 6 (1--5 scale)}
  \item{N7}{Neuroticism item 7 (1--5 scale)}
  \item{N8}{Neuroticism item 8 (1--5 scale)}
  \item{N9}{Neuroticism item 9 (1--5 scale)}
  \item{N10}{Neuroticism item 10 (1--5 scale)}
  \item{N_mean}{Mean of Neuroticism items}
  \item{O1}{Openness item 1 (1--5 scale)}
  \item{O2}{Openness item 2 (1--5 scale)}
  \item{O3}{Openness item 3 (1--5 scale)}
  \item{O4}{Openness item 4 (1--5 scale)}
  \item{O5}{Openness item 5 (1--5 scale)}
  \item{O6}{Openness item 6 (1--5 scale)}
  \item{O7}{Openness item 7 (1--5 scale)}
  \item{O8}{Openness item 8 (1--5 scale)}
  \item{O9}{Openness item 9 (1--5 scale)}
  \item{O10}{Openness item 10 (1--5 scale)}
  \item{O_mean}{Mean of Openness items}
}
}
\source{
Open-Source Psychometrics Project,
  \url{https://openpsychometrics.org/}
}
\usage{
BFI_Extraversion
}
\description{
Full responses to the 50-item Big Five Inventory (BFI; John & Srivastava,
1999) for N = 500 adults, rated on a 5-point Likert scale from 1
(Disagree strongly) to 5 (Agree strongly), along with demographic
variables. The ten Extraversion items (E1--E10) are used as the primary
example for the ordinal ALSI pipeline
(\code{\link{alsi_workflow_ordinal}}).
}
\references{
John, O. P., & Srivastava, S. (1999). The Big-Five trait taxonomy:
  History, measurement, and theoretical perspectives. In L. A. Pervin
  & O. P. John (Eds.), \emph{Handbook of personality: Theory and
  research} (2nd ed., pp. 102--138). Guilford Press.
}
\keyword{datasets}
