% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calsi-continuous.R, R/calsi.R
\name{calsi}
\alias{calsi}
\title{Compute Continuous Aggregated Latent Space Index (cALSI)}
\usage{
calsi(F, eig, K)

calsi(F, eig, K)
}
\arguments{
\item{F}{Matrix of row coordinates (N x K or larger)}

\item{eig}{Vector of eigenvalues}

\item{K}{Integer, number of dimensions to aggregate}
}
\value{
S3 object of class \code{calsi} containing:
  \item{alpha}{Numeric vector of cALSI values (length N)}
  \item{w}{Variance weights (length K)}
  \item{alpha_vec}{Aggregated direction vector (sqrt of weights)}
  \item{K}{Number of dimensions used}

S3 object of class \code{calsi} containing:
  \item{alpha}{Numeric vector of cALSI values (length N)}
  \item{w}{Variance weights (length K)}
  \item{alpha_vec}{Aggregated direction vector (sqrt of weights)}
  \item{K}{Number of dimensions used}
}
\description{
Calculates cALSI as a variance-weighted Euclidean norm of row coordinates
within a retained K-dimensional ipsatized SVD subspace.

Calculates cALSI as a variance-weighted Euclidean norm of row coordinates
within a retained K-dimensional ipsatized SVD subspace.
}
