% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{.alsi_make_disjunctive}
\alias{.alsi_make_disjunctive}
\title{Create disjunctive (indicator) matrix from binary data}
\usage{
.alsi_make_disjunctive(X01)
}
\description{
Create disjunctive (indicator) matrix from binary data
}
\keyword{internal}
