% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.check_class}
\alias{.check_class}
\title{Check if object inherits from expected class}
\usage{
.check_class(x, class_name, arg_name = "object")
}
\arguments{
\item{x}{Object to check}

\item{class_name}{Expected class name}

\item{arg_name}{Argument name for error message}
}
\description{
Check if object inherits from expected class
}
\keyword{internal}
