% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alsi.R, R/utilities.R
\name{.read_xlsx}
\alias{.read_xlsx}
\title{Read Excel file with fallback options}
\usage{
.read_xlsx(path)

.read_xlsx(path)
}
\arguments{
\item{path}{Path to .xlsx file}
}
\value{
Data frame
}
\description{
Read Excel file with fallback options

Read Excel file with fallback options
}
\keyword{internal}
