% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.validate_matrix}
\alias{.validate_matrix}
\title{Validate numeric matrix dimensions}
\usage{
.validate_matrix(X, min_rows = 2, min_cols = 2)
}
\arguments{
\item{X}{Matrix to validate}

\item{min_rows}{Minimum number of rows required}

\item{min_cols}{Minimum number of columns required}
}
\description{
Validate numeric matrix dimensions
}
\keyword{internal}
