% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alsi.R
\name{make_disjunctive}
\alias{make_disjunctive}
\title{Create disjunctive (indicator) matrix from binary data}
\usage{
make_disjunctive(X01)
}
\description{
Create disjunctive (indicator) matrix from binary data
}
\keyword{internal}
