% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alsi-categorical.R, R/alsi.R
\name{mca_indicator}
\alias{mca_indicator}
\title{Perform Multiple Correspondence Analysis on binary indicator matrix}
\usage{
mca_indicator(Xbin01)

mca_indicator(Xbin01)
}
\arguments{
\item{Xbin01}{Data frame or matrix with binary (0/1) variables}
}
\value{
List containing MCA results with eigenvalues, coordinates, and masses

List containing MCA results with eigenvalues, coordinates, and masses
}
\description{
Perform Multiple Correspondence Analysis on binary indicator matrix

Perform Multiple Correspondence Analysis on binary indicator matrix
}
\keyword{internal}
