% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calsi-continuous.R, R/calsi.R
\name{plot_domain_loadings}
\alias{plot_domain_loadings}
\title{Plot Domain Loadings in SVD Space}
\usage{
plot_domain_loadings(fit, dim_pair = c(1, 2), cex = 1)

plot_domain_loadings(fit, dim_pair = c(1, 2), cex = 1)
}
\arguments{
\item{fit}{SVD fit object (class \code{svd_fit})}

\item{dim_pair}{Integer vector of length 2, dimensions to plot}

\item{cex}{Character expansion for labels}
}
\description{
Visualizes domain loadings in a 2D subspace (biplot-style).

Visualizes domain loadings in a 2D subspace (biplot-style).
}
