% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calsi-continuous.R, R/calsi.R
\name{plot_subspace_stability_cont}
\alias{plot_subspace_stability_cont}
\title{Plot Subspace Stability Diagnostics for Continuous Data}
\usage{
plot_subspace_stability_cont(boot_obj)

plot_subspace_stability_cont(boot_obj)
}
\arguments{
\item{boot_obj}{Object of class \code{svd_bootstrap}}
}
\description{
Plot Subspace Stability Diagnostics for Continuous Data

Plot Subspace Stability Diagnostics for Continuous Data
}
