% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calsi.R
\name{principal_angles_rad}
\alias{principal_angles_rad}
\title{Principal angles between column spaces of two matrices}
\usage{
principal_angles_rad(A, B)
}
\description{
Principal angles between column spaces of two matrices
}
\keyword{internal}
