% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calsi.R
\name{procrustes_rotation}
\alias{procrustes_rotation}
\title{Procrustes rotation to align two matrices}
\usage{
procrustes_rotation(X, Y)
}
\description{
Procrustes rotation to align two matrices
}
\keyword{internal}
