% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alsi.R, R/calsi.R, R/utilities.R
\name{summarise_matrix}
\alias{summarise_matrix}
\title{Summarize matrix columns (median and quantiles)}
\usage{
summarise_matrix(X, probs = c(0.05, 0.95))

summarise_matrix(X, probs = c(0.05, 0.95))

summarise_matrix(X, probs = c(0.05, 0.95))
}
\arguments{
\item{X}{Matrix to summarize}

\item{probs}{Numeric vector of probabilities for quantiles}
}
\value{
Matrix with median and quantile summaries for each column
}
\description{
Summarize matrix columns (median and quantiles)

Summarize matrix columns (median and quantiles)

Summarize matrix columns (median and quantiles)
}
\keyword{internal}
