% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{svd_align}
\alias{svd_align}
\title{Align SVD solution via Procrustes rotation with sign anchoring}
\usage{
svd_align(B, Bref)
}
\arguments{
\item{B}{Matrix of domain loadings to align}

\item{Bref}{Reference matrix of domain loadings}
}
\value{
List with aligned coordinates and rotation matrix
}
\description{
Align SVD solution via Procrustes rotation with sign anchoring
}
