% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calsi-continuous.R, R/calsi.R
\name{svd_ipsatized}
\alias{svd_ipsatized}
\title{Perform ipsatized SVD on continuous data (SEPA-style)}
\usage{
svd_ipsatized(X, K = NULL)

svd_ipsatized(X, K = NULL)
}
\arguments{
\item{X}{Data frame or matrix with continuous variables (persons x domains)}

\item{K}{Number of dimensions to retain (default: all)}
}
\value{
List containing SVD results with eigenvalues, coordinates

List containing SVD results with eigenvalues, coordinates
}
\description{
Perform ipsatized SVD on continuous data (SEPA-style)

Perform ipsatized SVD on continuous data (SEPA-style)
}
\keyword{internal}
