% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alsi-categorical.R, R/alsi.R
\name{to01}
\alias{to01}
\title{Read Excel file with fallback options}
\usage{
to01(x)

to01(x)
}
\description{
Read Excel file with fallback options

Convert various formats to binary 0/1
}
\keyword{0/1}
\keyword{Convert}
\keyword{binary}
\keyword{formats}
\keyword{internal}
\keyword{to}
\keyword{various}
