% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alsi.R
\name{summarise_matrix}
\alias{summarise_matrix}
\title{Summarize matrix columns (median and quantiles)}
\usage{
summarise_matrix(X, probs = c(0.05, 0.95))
}
\description{
Summarize matrix columns (median and quantiles)
}
\keyword{internal}
